Minor adjustments

+++ server/controllers/exam.controller.js
@@ -14,6 +14,11 @@ class ExamController {
   static async getExamById(req, res) {
     try {
       const { id } = req.params;
+
+      if (!id.match(/^[0-9a-fA-F]{24}$/)) {
+        return res.status(400).json({ error: 'Invalid exam ID' });
+      }
+
       const exam = await ExamModel.findById(id);
       if (!exam) {
         return res.status(404).json({ error: 'Exam not found' });
@@ -57,6 +62,10 @@ class ExamController {
       const { id } = req.params;
       const { name, questions, timeLimit } = req.body;
 
+      if (!id.match(/^[0-9a-fA-F]{24}$/)) {
+        return res.status(400).json({ error: 'Invalid exam ID' });
+      }
+
       for (const question of questions) {
         const trueAnswerCount = question.answers.filter(
           (answer) => answer.isCorrect
@@ -92,6 +101,11 @@ class ExamController {
   static async deleteExam(req, res) {
     try {
       const { id } = req.params;
+
+      if (!id.match(/^[0-9a-fA-F]{24}$/)) {
+        return res.status(400).json({ error: 'Invalid exam ID' });
+      }
+
       const deletedExam = await ExamModel.findByIdAndDelete(id);
       if (!deletedExam) {
         return res.status(404).json({ error: 'Exam not found' });
